fix: return metadata for Goldfinch PoolTokens
Goldfinch PoolTokens don't return metadata because they aren't correctly indexed by
OpenSea as ERC721s. This is likely because the contract did not return true for
`supportsInterface(INTERFACE_ID_ERC721)` at the time of contract creation, which
appears to be a criteria for being detected as an ERC721 (this has since been
fixed). Work around the issue by querying the metadata URI directly.